+++ tests_mypy/enums.py
@@ -1,6 +1,6 @@
 from enum import Enum, IntEnum
 
-from ariadne import EnumType
+from ariadne import EnumType, make_executable_schema
 
 
 class TestEnum(Enum):
@@ -16,3 +16,18 @@ class TestIntEnum(IntEnum):
 test_dict_enum_type = EnumType("TestDictEnum", {"FOO": "foo", "BAR": "bar"})
 test_enum_type = EnumType("TestEnum", TestEnum)
 test_int_enum_type = EnumType("TestIntEnumType", TestIntEnum)
+
+
+make_executable_schema(
+    """
+    type Query {
+        test: TestEnum!
+    }
+
+    enum TestEnum {
+        FOO
+        VAR
+    }
+    """,
+    TestEnum,
+)